Add self-expire tests for setting field
